Organize host_virtual_switch and host_port_group structure helpers
There are a lot of helpers here, and I want to make sure that they are
at least somewhat organized and annotated. Hence I've split them up in a
way I think will make sense.
