+++ vsphere/resource_vsphere_host_port_group.go
@@ -7,10 +7,7 @@ import (
 	"context"
 
 	"github.com/hashicorp/terraform/helper/schema"
-	"github.com/hashicorp/terraform/terraform"
 	"github.com/vmware/govmomi"
-	"github.com/vmware/govmomi/vim25/mo"
-	"github.com/vmware/govmomi/vim25/types"
 )
 
 func resourceVSphereHostPortGroup() *schema.Resource {
@@ -147,85 +144,4 @@ func resourceVSphereHostPortGroupDelete(d *schema.ResourceData, meta interface{}
 
 	d.SetId("")
 	return nil
-}
-
-// hostPortGroupFromName locates a host port group from its assigned name and
-// host, using the client's default property collector.
-func hostPortGroupFromName(client *govmomi.Client, name, host, datacenter string, timeout time.Duration) (*types.HostPortGroup, error) {
-	ns, err := hostNetworkSystemFromName(client, host, datacenter)
-	if err != nil {
-		return nil, fmt.Errorf("error loading network system: %s", err)
-	}
-
-	var mns mo.HostNetworkSystem
-	pc := client.PropertyCollector()
-	ctx, cancel := context.WithTimeout(context.Background(), timeout)
-	defer cancel()
-	if err := pc.RetrieveOne(ctx, ns.Reference(), []string{"networkInfo.portgroup"}, &mns); err != nil {
-		return nil, fmt.Errorf("error fetching host network properties: %s", err)
-	}
-
-	for _, pg := range mns.NetworkInfo.Portgroup {
-		if pg.Spec.Name == name {
-			return &pg, nil
-		}
-	}
-
-	return nil, fmt.Errorf("port group %s not found on host %s", name, host)
-}
-
-// calculateComputedPolicy is a utility function to compute a map of state
-// attributes for the port group's effective policy. It uses a bit of a
-// roundabout way to set the attributes, but allows us to utilize our
-// functional deep reading helpers to perform this task, versus having to
-// re-write code.
-//
-// This function relies a bit on some of the lower-level utility functionality
-// of helper/schema, so it may need to change in the future.
-func calculateComputedPolicy(policy types.HostNetworkPolicy) (map[string]string, error) {
-	cpr := &schema.Resource{Schema: schemaHostNetworkPolicy()}
-	cpd := cpr.Data(&terraform.InstanceState{})
-	cpd.SetId("effectivepolicy")
-	if err := flattenHostNetworkPolicy(cpd, &policy); err != nil {
-		return nil, fmt.Errorf("error setting effective policy data: %s", err)
-	}
-	cpm := cpd.State().Attributes
-	delete(cpm, "id")
-	return cpm, nil
-}
-
-// calculatePorts is a utility function that returns a set of port data.
-func calculatePorts(ports []types.HostPortGroupPort) *schema.Set {
-	s := make([]interface{}, 0)
-	for _, port := range ports {
-		m := make(map[string]interface{})
-		m["key"] = port.Key
-		m["mac_addresses"] = sliceStringsToInterfaces(port.Mac)
-		m["type"] = port.Type
-		s = append(s, m)
-	}
-	return schema.NewSet(schema.HashResource(portGroupPortSchema()), s)
-}
-
-func portGroupPortSchema() *schema.Resource {
-	return &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"key": &schema.Schema{
-				Type:        schema.TypeString,
-				Description: "The linkable identifier for this port entry.",
-				Computed:    true,
-			},
-			"mac_addresses": &schema.Schema{
-				Type:        schema.TypeList,
-				Description: "The MAC addresses of the network service of the virtual machine connected on this port.",
-				Computed:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"type": &schema.Schema{
-				Type:        schema.TypeString,
-				Description: "Type type of the entity connected on this port. Possible values are host (VMKkernel), systemManagement (service console), virtualMachine, or unknown.",
-				Computed:    true,
-			},
-		},
-	}
 }
